moved plotting fns, GPs fit for mean, fixed tests to reflect changes
