+++ approxposterior/__init__.py
@@ -18,3 +18,4 @@ from .utility import *
 from .likelihood import *
 from .gp_utils import *
 from .mcmc_utils import *
+from .plot_utils import *
